Make the readme more inclusive
It's better to use gender neutral terminology in documentation, otherwise it can contribute to people feeling unwelcome in the open source community, or in software engineering as a discipline. 

Also, some of the language used came across to me as a little patronising for beginners. I understand English may not be your first language, and I must say it's great documentation in general given that, but I think these few changes make it a little more approachable to newcomers or those less experienced with Python.