Handle upgrades of packages within ranges
The underlying functionality for this was already present within
`pip-tool`s, however `pip-compile-multi` wasn't aware of it and
thus made assumptions about the nature of the spec argument which
turned out to be false.

The fix here is to teach `pip-compile-multi` how to interpret these
package specs to extract the names for the places it needs them
while still passing the full spaces through to `pip-tools` for the
actual upgrade.

Fixes https://github.com/peterdemin/pip-compile-multi/issues/392