Extract util to simplify using test data without locally modifying it
This makes it easier to use test data stored in the repo without
the tests actually modifying the repo content. While this makes it
slightly harder to see the changes that a test failure made, it
avoids creating unexpected changes to the developer's working
environment when they run the tests.